Conflicts:
	nts.uk/uk.com/build.gradle

+++ nts.uk/uk.com/com.ctx/sys/portal/nts.uk.ctx.sys.portal.app/src/main/java/nts/uk/ctx/sys/portal/app/command/webmenu/RemoveWebMenuCommandHander.java
@@ -1,12 +1,18 @@
 package nts.uk.ctx.sys.portal.app.command.webmenu;
 
+import java.util.Optional;
+
 import javax.ejb.Stateless;
 import javax.inject.Inject;
 import javax.transaction.Transactional;
 
+import nts.arc.error.BusinessException;
 import nts.arc.layer.app.command.CommandHandler;
 import nts.arc.layer.app.command.CommandHandlerContext;
+import nts.uk.ctx.sys.portal.dom.webmenu.DefaultMenu;
+import nts.uk.ctx.sys.portal.dom.webmenu.WebMenu;
 import nts.uk.ctx.sys.portal.dom.webmenu.WebMenuRepository;
+import nts.uk.ctx.sys.portal.dom.webmenu.jobtitletying.JobTitleTyingRepository;
 import nts.uk.shr.com.context.AppContexts;
 
 @Stateless
@@ -15,6 +21,9 @@ public class RemoveWebMenuCommandHander extends CommandHandler<RemoveWebMenuComm
 	
 	@Inject
 	private WebMenuRepository webMenuRepository;
+	
+	@Inject
+	private JobTitleTyingRepository jobTitleTyingRepository;
 
 	@Override
 	protected void handle(CommandHandlerContext<RemoveWebMenuCommand> context) {
@@ -22,7 +31,18 @@ public class RemoveWebMenuCommandHander extends CommandHandler<RemoveWebMenuComm
 		RemoveWebMenuCommand command = context.getCommand();
 		String companyId = AppContexts.user().companyId();
 		
+		Optional<WebMenu> webMenu = webMenuRepository.find(companyId, command.getWebMenuCd());
+		if (!webMenu.isPresent()) {
+			throw new RuntimeException("Not found web menu code:" + command.getWebMenuCd());
+		}
+		
+		if (webMenu.get().isDefault()) {
+			throw new BusinessException("Msg_72");
+		}
+		
 		webMenuRepository.remove(companyId, command.getWebMenuCd());
+		
+		jobTitleTyingRepository.removeByMenuCode(companyId, command.getWebMenuCd());
 	}
